+++ src/resource/RefraxResourceBase.js
@@ -8,6 +8,7 @@
 const Promise = require('bluebird');
 const mixinSubscribable = require('mixinSubscribable');
 const mixinConfigurable = require('mixinConfigurable');
+const mixinDisposable = require('mixinDisposable');
 const RefraxOptions = require('RefraxOptions');
 const RefraxParameters = require('RefraxParameters');
 const RefraxQueryParameters = require('RefraxQueryParameters');
@@ -68,7 +69,8 @@ class RefraxResourceBase {
     Object.defineProperty(this, '_accessor', {value: accessor});
     Object.defineProperty(this, '_paths', {value: paths});
 
-    mixinSubscribable(this);
+    mixinDisposable(this);
+    this.onDispose(mixinSubscribable.asDisposable(this));
     mixinConfigurable(this, {
       _options: options,
       _parameters: parameters,