+++ src/util/mixinSubscribable.js
@@ -41,6 +41,36 @@ const Mixin = {
     };
   },
 
+  once: function(event, callback, context) {
+    var disposed = false
+      , eventHandler = null;
+
+    if (typeof(event) !== 'string') {
+      throw new TypeError('mixinSubscribable - subscribe expected string event but found type `' + event + '`');
+    }
+
+    if (typeof(callback) !== 'function') {
+      throw new TypeError('mixinSubscribable - subscribe expected callback but found `' + event + '`');
+    }
+
+    context = context || this;
+    eventHandler = function() {
+      if (disposed) {
+        return;
+      }
+      disposed = true;
+      callback.apply(context, arguments);
+    };
+
+    this._emitter.once(event, eventHandler, context || this);
+    return () => {
+      if (!disposed) {
+        disposed = true;
+        this._emitter.removeListener(event, eventHandler);
+      }
+    };
+  },
+
   emit: function() {
     return this._emitter.emit.apply(this._emitter, arguments);
   }
@@ -57,5 +87,13 @@ function mixinSubscribable(target) {
 
   return RefraxTools.extend(target, Mixin);
 }
+
+mixinSubscribable.asDisposable = function(target) {
+  mixinSubscribable(target);
+
+  return () => {
+    target._emitter.removeAllListeners();
+  };
+};
 
 export default mixinSubscribable;